+++ src/zip/zfile.h
@@ -25,11 +25,11 @@ void zip_close(void);
 int zip_findfirst(struct zip_find_t *file);
 int zip_findnext(struct zip_find_t *file);
 
-int zopen(const char *filename);
-size_t zread(int fd, void *buf, size_t size);
-int zgetc(int fd);
-int zclose(int fd);
-size_t zsize(int fd);
+int64_t zopen(const char *filename);
+size_t zread(int64_t fd, void *buf, size_t size);
+int zgetc(int64_t fd);
+int zclose(int64_t fd);
+size_t zsize(int64_t fd);
 #if (EMU_SYSTEM == NCDZ)
 int zlength(const char *filename);
 #endif
